fix(desktop): use __mocks__ for walletconnect sign-client in tests
With the static import of @walletconnect/sign-client in
coin-concordium, jest.mock with virtual: true no longer
intercepts requires from within the built coin module.
Auto-mock via __mocks__ directory resolves globally.
